add second_degree.c

+++ includes/computor.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   computor.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nsalle <nsalle@student.42lyon.fr>          +#+  +:+       +#+        */
+/*   By: rcodazzi <rcodazzi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/14 13:39:36 by nsalle            #+#    #+#             */
-/*   Updated: 2021/01/15 12:27:33 by nsalle           ###   ########lyon.fr   */
+/*   Updated: 2021/01/15 15:25:24 by rcodazzi         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,18 +18,18 @@
 typedef struct		s_expr
 {
     char *expr;
-    int degree1;
-    int degree2;
-    int degree3;
-    int degree0;
-    int discriminant;
+    float degree1;
+    float degree2;
+    float degree3;
+    float degree0;
+    float discriminant;
 }					t_expr;
 
 typedef struct		s_maths
 {
     t_expr  *left_exp;
     t_expr	*right_exp;
-    int		degree;
+    float	degree;
 }					t_maths;
 
 void    split_expressions(t_maths *master, char *argv);
